+++ lib/scarpe/wv/drawable.rb
@@ -153,7 +153,7 @@ module Scarpe::Webview
     #
     # @return [Scarpe::WebWrangler::ElementWrangler] a DOM object manager
     def html_element
-      @elt_wrangler ||= Scarpe::Webview::WebWrangler::ElementWrangler.new(html_id)
+      @elt_wrangler ||= Scarpe::Webview::WebWrangler::ElementWrangler.new(html_id:)
     end
 
     # Return a promise that guarantees all currently-requested changes have completed